netifd: trigger action event after interface-down, and autostart set
Previously, the netifd action event was not being triggered when an
interface-down event happened on an interface with autostart set: the
autostart flag caused the interface to move to the setup state, and so
the action event was not being triggered.

Now, when an interface-down event happens, instead of testing the
current state, we test the previous state (before it was possibly
changed), which appears to be the intent of the original commit, here:

    3e7807e93cea01c779bd112a25a8b166bc9fdf8c

An interface that fails to be brought up still (correctly) does not
generate an action event because it was previously in the setup state.

DAL-6364