user/netifd: increased teardown timeout
Cancelled teardowns are undesirable. Err on the side of waiting a long time.

The increased teardown timeout is particularly useful for mm proto teardowns
that can call mmcli commands that timeout on poorly behaved modems.
Interrupting the teardown causes dynamic child interfaces to hang around
in the "down" state even if the parent interface is removed.
Once "down" the child interfaces remain "down" even when the parent interface
is re-added and tries to use the children.

Signed-off-by: Anubhav Gupta <anubhav.gupta@digi.com>