Modem reconnect: Ensure the modem interface doesn't disappear when toggled
If the modem is enabled and disabled in quick succession, the interface
can get deleted. This happens in netifd if the interface hasn't completed
its removal before being re-added. What then happens is the interface is
updated before the removal is complete - and is removed when the process
completes. From this point onwards, netifd won't respond to the interface
and for all intents and purposes, it is unavailable.

By detecting this condition and stopping the removal process, the updated
interface remains up and functional. There are three cases when modifying
an interface in netifd:

creation
removal
update

Creation and removal work as expected, however if we get an update (ie
creating whilst removing) then we can get into the error state. When
we do an update and the old interface is marked as for removal, then
stop that removal by setting its connect state to NORMAL. This stops the
interface being removed at a later time.

See JIRA:DAL-1432

Signed-off-by: Anubhav Gupta <anubhav.gupta@digi.com>