user/netifd: set config reload flag in proto setup
Set an environment variable when the proto shell script is being run
because of a "reload" event.
This can be observed in the proto "setup" after a configuration change
is made to the interface.
When the variable is not present, one can assume that an down/up event
is responsible for the setup.

Determining reload vs down/up is required for the passthrough.sh proto
to decide whether to toggle link on the underlying network device.

DAL-2642

Signed-off-by: Anubhav Gupta <anubhav.gupta@digi.com>