user/netifd: install IPv6 routes with prefsrc instead of src
Previously, any IPv6 route pushed into netifd with a source address
silently failed to be inserted into the kernel because the RTA_SRC
netlink route attribute requires the Linux IPV6_SUBTREES feature to
work.

This became an issue for IPv6 request-a-network (dynamic IP) IPsec policies
where the assigned IP address was being used for the route source
address, and failing to be inserted into the kernel.

The RTA_SRC attribute (Route source address) is not actually what we
want anyway. We want the route to have the RTA_PREFSRC (Preferred source
address) set, so an appropriate source address is chosen for locally
generated packets. Just like for IPv4.

This change makes netifd's IPv6 route source attribute match the use of
that same attribute with IPv4.

DAL-3910

Signed-off-by: Anubhav Gupta <anubhav.gupta@digi.com>