treewide: simplify use of uci_blob_diff
Simplify use of uci_blob_diff by dropping the use of the diff variable.
The diff variable is not used and the current usage of uci_blob_diff is
to only know if there are difference in the 2 blob, we are not
interested in what are the actual difference. Hence the diff can be
dropped.

With the diff dropped, we can drop the compile check there to make sure
__DEV_ATTR_MAX was small enough to be in the size of diff (unsigned
long). This is needed in preparation of adding additional ATTR that
would make __DEV_ATTR_MAX exceed the size of the diff variable.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>